refactor(runs): split PR-merge/URL-parse group into pr-checks.ts
Move checkPullRequestMerged, mergePullRequest, parsePullRequestUrl,
parsePullRequestRef, isRateLimited and their exported types/regexes out
of src/runs/pr.ts into a new src/runs/pr-checks.ts. Shared GitHub REST
helpers (buildHeaders/readJson/readText/truncate/readGhFetchOverride/
GITHUB_API_BASE/GH_FETCH_OVERRIDE_ENV) live in pr-checks.ts and are
imported by pr.ts (one copy, no duplication). pr.ts re-exports every
moved public symbol so the ../runs/pr.ts import path still resolves.

pr.ts drops from 659 to 361 lines. mergePullRequest's grandfathered
complexity override follows it to pr-checks.ts in biome.json.

Refs: warren-db9a (pl-88bb step 1)
