+++ .canopy/agent.json
@@ -14,6 +14,7 @@
     ],
     "model": "claude-opus-4-8",
     "source": "builtin",
+    "runtime": "pi",
     "provider": "anthropic"
   }
 }
